+++ plone/app/contenttypes/indexers.py
@@ -42,10 +42,13 @@ def SearchableText(obj):
         textvalue = richtext.text
         if IRichTextValue.providedBy(textvalue):
             transforms = getToolByName(obj, 'portal_transforms')
+            # Before you think about switching raw/output
+            # or mimeType/outputMimeType, first read
+            # https://github.com/plone/Products.CMFPlone/issues/2066
             text = transforms.convertTo(
                 'text/plain',
-                safe_unicode(textvalue.output).encode('utf8'),
-                mimetype=textvalue.outputMimeType,
+                safe_unicode(textvalue.raw).encode('utf-8'),
+                mimetype=textvalue.mimeType,
             ).getData().strip()
 
     subject = u' '.join(